requirements.txt
pandas
numpy
matplotlib
pillow
torch
torchvision
scikit-learn
opencv-python
transformers
